requirements.txt
# Core dependencies
jax>=0.4.0
jaxlib>=0.4.0
jax-md>=0.2.0
numpy>=1.24.0

# Visualization
pygame>=2.5.0
matplotlib>=3.7.0
imageio>=2.31.0

# ML and clustering
scikit-learn>=1.3.0

# Experiment tracking
wandb>=0.15.0

# Progress bars
tqdm>=4.65.0

# Configuration management
hydra-core>=1.3.0
omegaconf>=2.3.0

# Optional: For GPU support, install CUDA-enabled jaxlib
# Follow instructions at: https://github.com/google/jax#installation
